Extract download action to new separeted class
